requirements.txt
aiofiles==22.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.0
aiosignal==1.3.2
alembic==1.14.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1731428368116/work
croniter==1.3.7
cryptography==42.0.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
dask==2025.1.0
datadog==0.44.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
Faker==4.14.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1738203752110/work
fqdn==1.5.1
freezegun==1.2.2
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.41
great-expectations==0.18.12
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.15.0
ipython==8.10.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==7.4.4
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.0.13
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1725459273855/work
ldap3==2.9.1
Levenshtein==0.26.1
locket==1.0.0
mage-ai==0.9.75
makefun==1.15.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.1
multidict==6.1.0
munkres==1.1.4
mysql-connector-python @ file:///Users/runner/miniforge3/conda-bld/mysql-connector-python_1725367376714/work/mysql-connector-python
narwhals @ file:///home/conda/feedstock_root/build_artifacts/narwhals_1738439166728/work
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
newrelic==8.8.0
notebook==7.0.8
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1726878435917/work
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.3.6
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1738287743623/work
polars==0.19.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==1.59.0
Pygments==2.19.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1735698276679/work
python-dateutil==2.8.2
python-json-logger==3.2.1
python-Levenshtein==0.26.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
redis==5.0.8
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.17.17
scikit-learn==1.6.1
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1736617280897/work/dist/scipy-1.15.1-cp310-cp310-macosx_11_0_arm64.whl#sha256=5f75e798ec682bba92f312fbafcfd7036b41f1516712c8f9ca6214156476bd8d
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
Send2Trash==1.8.3
sentry-sdk==1.19.1
shellingham==1.5.4
simplejson==3.19.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
sqlglot==26.3.9
sqlglotrs==0.3.14
stack-data==0.6.3
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1727986736668/work
terminado==0.17.1
text-unidecode==1.3
thefuzz==0.19.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.9.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
tzlocal==5.2
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1736692581912/work
uri-template==1.3.0
urllib3==2.3.0
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
